DO NOT MERGE: expansion of dwalsh's comparison script
Something that compares _all_ subcommands. It's a horrible
mess because some man pages display COMMANDS and some
display SUBCOMMANDS; and some show the subcommand name
in the first column, others show the podman-xxx man page;
and there are inconsistencies galore.

This is a fake PR so Dan can look at the rest of the
inconsistencies that his script didn't look at

Signed-off-by: Ed Santiago <santiago@redhat.com>